requirements.txt
python-rake
numpy
pyyaml
nltk
flask
